(WiiU) Fix for modern devkitPPC versions
This change removes the old -mwup flag that is no longer in current versions of devkitPPC. The equivalent flags are provided instead.

This change also removes the int/long int tweaks required for the version of devkitPPC that RetroArch was pinned to. This will cause new, non-fatal, meaningless compiler warnings on the old version. The build and core still work fine. These warnings will be silenced when the project updates to a newer version of devkitPPC.